@@ -11,35 +11,37 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^16.2.8",
-    "@angular/common": "^16.2.8",
-    "@angular/compiler": "^16.2.8",
-    "@angular/core": "^16.2.8",
-    "@angular/forms": "^16.2.8",
-    "@angular/platform-browser": "^16.2.8",
-    "@angular/platform-browser-dynamic": "^16.2.8",
-    "@stomp/stompjs": "^7.0.0",
-    "primeicons": "^6.0.1",
-    "primeng": "^16.2.0",
-    "rxjs": "~7.8.0",
-    "tslib": "^2.3.0",
-    "ws": "^8.14.2",
-    "ytdl-core": "^4.11.5",
-    "zone.js": "~0.13.0"
+    "@angular/animations": "17.1.2",
+    "@angular/common": "17.1.2",
+    "@angular/compiler": "17.1.2",
+    "@angular/core": "17.1.2",
+    "@angular/forms": "17.1.2",
+    "@angular/platform-browser": "17.1.2",
+    "@angular/platform-browser-dynamic": "17.1.2",
+    "@angular/router": "17.1.2",
+    "@auth0/angular-jwt": "5.2.0",
+    "@stomp/stompjs": "7.0.0",
+    "primeicons": "6.0.1",
+    "primeng": "17.2",
+    "rxjs": "7.8.1",
+    "tslib": "2.3.0",
+    "vite": "4.5.2",
+    "ws": "8.14.2",
+    "ytdl-core": "4.11.5",
+    "zone.js": "0.14.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^16.2.8",
-    "@angular/cli": "~16.2.8",
-    "@angular/compiler-cli": "^16.2.8",
-    "@angular/router": "^16.2.8",
-    "@types/jasmine": "~4.3.0",
-    "eslint": "^8.47.0",
-    "jasmine-core": "~4.6.0",
-    "karma": "~6.4.0",
-    "karma-chrome-launcher": "~3.2.0",
-    "karma-coverage": "~2.2.0",
-    "karma-jasmine": "~5.1.0",
-    "karma-jasmine-html-reporter": "~2.1.0",
-    "typescript": "^5.1.6"
+    "@angular-devkit/build-angular": "17.1.2",
+    "@angular/cli": "17.1.2",
+    "@angular/compiler-cli": "17.1.2",
+    "@types/jasmine": "4.3.0",
+    "eslint": "8.47.0",
+    "jasmine-core": "4.6.0",
+    "karma": "6.4.0",
+    "karma-chrome-launcher": "3.2.0",
+    "karma-coverage": "2.2.0",
+    "karma-jasmine": "5.1.0",
+    "karma-jasmine-html-reporter": "2.1.0",
+    "typescript": "5.2.2"
   }
 }
