@@ -1,15 +1,58 @@
 package ly.happylone.config;
 
 import org.springframework.context.annotation.Configuration;
+import org.springframework.http.server.ServerHttpRequest;
+import org.springframework.lang.NonNull;
+import org.springframework.messaging.Message;
+import org.springframework.messaging.MessageChannel;
+import org.springframework.messaging.simp.config.ChannelRegistration;
 import org.springframework.messaging.simp.config.MessageBrokerRegistry;
+import org.springframework.messaging.simp.stomp.StompHeaderAccessor;
+import org.springframework.messaging.support.ChannelInterceptor;
+import org.springframework.messaging.support.MessageBuilder;
+import org.springframework.security.core.Authentication;
+import org.springframework.web.socket.WebSocketHandler;
 import org.springframework.web.socket.config.annotation.EnableWebSocketMessageBroker;
 import org.springframework.web.socket.config.annotation.StompEndpointRegistry;
 import org.springframework.web.socket.config.annotation.WebSocketMessageBrokerConfigurer;
-import org.springframework.lang.NonNull;
+import org.springframework.web.socket.server.support.DefaultHandshakeHandler;
+
+import java.security.Principal;
+import java.util.Map;
 
 @Configuration
 @EnableWebSocketMessageBroker
 public class WebSocketConfig implements WebSocketMessageBrokerConfigurer {
+
+    private final JWTAuthenticationFilter jwtAuthenticationFilter;
+
+    public WebSocketConfig(JWTAuthenticationFilter jwtAuthenticationFilter) {
+        this.jwtAuthenticationFilter = jwtAuthenticationFilter;
+    }
+
+    @Override
+    public void configureClientInboundChannel(@NonNull ChannelRegistration registration) {
+        registration.interceptors(new ChannelInterceptor() {
+            @Override
+            public Message<?> preSend(@NonNull Message<?> message, @NonNull MessageChannel channel) {
+                StompHeaderAccessor accessor = StompHeaderAccessor.wrap(message);
+
+                // Extract the token from the query parameters
+                String token = accessor.getFirstNativeHeader("token");
+
+                // validate and convert to a Principal based on your own requirements e.g.
+                // authenticationManager.authenticate(JwtAuthentication(token))
+                Principal yourAuth = token == null ? null : jwtAuthenticationFilter.getAuthentication(token);
+
+                accessor.setUser(yourAuth);
+
+                // not documented anywhere but necessary otherwise NPE in
+                // StompSubProtocolHandler!
+                accessor.setLeaveMutable(true);
+                return MessageBuilder.createMessage(message.getPayload(), accessor.getMessageHeaders());
+            }
+        });
+    }
 
     @Override
     public void configureMessageBroker(@NonNull MessageBrokerRegistry config) {
@@ -20,7 +63,22 @@ public class WebSocketConfig implements WebSocketMessageBrokerConfigurer {
     @Override
     public void registerStompEndpoints(@NonNull StompEndpointRegistry registry) {
         registry.addEndpoint("/websocket")
-                .setAllowedOrigins()
+                .setAllowedOrigins("http://localhost:4200") // Replace with your origin
+                .setHandshakeHandler(new DefaultHandshakeHandler() {
+                    @Override
+                    protected Principal determineUser(@NonNull ServerHttpRequest request,
+                            @NonNull WebSocketHandler wsHandler,
+                            @NonNull Map<String, Object> attributes) {
+                        String token = request.getURI().getQuery().split("token=")[1].split("&")[0];
+                        if (token != null) {
+                            Authentication auth = jwtAuthenticationFilter.getAuthentication(token);
+                            if (auth != null && auth.isAuthenticated()) {
+                                return auth;
+                            }
+                        }
+                        return null;
+                    }
+                })
                 .withSockJS();
     }
 }
