@@ -11,7 +11,7 @@ import { ThreadCreateComponent } from './thread-create/thread-create.component';
 import { ThreadDetailComponent } from './thread-detail/thread-detail.component';
 import { LoginComponent } from './login/login.component';
 import { RegisterComponent } from './register/register.component';
-import { AuthGuard } from './auth.guard';
+import { authGuard } from './auth.guard';
 import { HeaderComponent } from './header/header.component';
 
 
@@ -19,34 +19,19 @@ const routes: Routes = [
   { path: '', redirectTo: 'login', pathMatch: 'full' },
   { path: 'login', component: LoginComponent },
   { path: 'register', component: RegisterComponent },
-  { path: 'home', component: HomeComponent, canActivate: [AuthGuard] },
-  { path: 'art', component: ArtComponent, canActivate: [AuthGuard] },
-  { path: 'shop', component: YoutubeDLComponent, canActivate: [AuthGuard] },
-  { path: 'shop', component: ShopComponent, canActivate: [AuthGuard] },
-  { path: 'user', component: UserComponent, canActivate: [AuthGuard] },
-  { path: 'admin', component: AdminComponent, canActivate: [AuthGuard] },
-  { path: 'youtube-dl', component: YoutubeDLComponent, canActivate: [AuthGuard] },
-  { path: 'threads', component: ThreadListComponent, canActivate: [AuthGuard] },
-  { path: 'create-thread', component: ThreadCreateComponent, canActivate: [AuthGuard] },
-  { path: 'thread/:id', component: ThreadDetailComponent, canActivate: [AuthGuard] },
-  { path: 'header', component: HeaderComponent, canActivate: [AuthGuard] },
+  { path: 'home', component: HomeComponent, canActivate: [authGuard] },
+  { path: 'art', component: ArtComponent, canActivate: [authGuard] },
+  { path: 'shop', component: ShopComponent, canActivate: [authGuard] },
+  { path: 'user', component: UserComponent, canActivate: [authGuard] },
+  { path: 'admin', component: AdminComponent, canActivate: [authGuard] },
+  { path: 'youtube-dl', component: YoutubeDLComponent, canActivate: [authGuard] },
+  { path: 'threads', component: ThreadListComponent, canActivate: [authGuard] },
+  { path: 'create-thread', component: ThreadCreateComponent, canActivate: [authGuard] },
+  { path: 'thread/:id', component: ThreadDetailComponent, canActivate: [authGuard] },
+  { path: 'header', component: HeaderComponent, canActivate: [authGuard] },
   { path: "**", redirectTo: "login" },
 ];
-/*
 
-const routes: Routes = [
-  { path: '', redirectTo: '/login', pathMatch: 'full' },
-  { path: 'register', component: RegistrationComponent },
-  { path: 'login', component: LoginComponent },
-  { path: '', redirectTo: '/home', pathMatch: 'full' },
-  { path: 'home', component: HomeComponent },
-  { path: 'art', component: ArtComponent },
-  { path: 'shop', component: ShopComponent },
-  { path: 'user', component: UserComponent },
-  { path: 'admin', component: AdminComponent },
-  { path: "**", redirectTo: "/home" }
-];
-*/
 @NgModule({
   imports: [RouterModule.forRoot(routes, { useHash: true })],
   exports: [RouterModule]
